Stub out a python csv parser.
We have our annotated text in a Google Doc Spreadsheet, and
we need to turn it into TEI-ish sgml for ingest into the
Coptic Scriptorium data set, as well as other formats
like Markdown or HTML for convenient reading.

This is the start of a quick script to parse the text and
other relevant fields out of a CSV (comma-separated values)
file exported from the spreadsheet.

Currently it just prints each line as plain text along with
our annotations of Drescher's line citation scheme