Add a column for comments.
We have some comments on how we understand certain parts of
the text. Try to display those as well.

We keep track of the longest Coptic line, and then hard-code
that as a column width. Unfortunately this doesn't actually
work for an even column boundary because the format expansion
counts bytes (or characters?) rather than display length.

Unicode text is still hard in python.