Fill in continuous line numbers.
The first column has periodic page.line references to Drescher's
edition of the text. Try to parse these, and use them as a
basis for generating a corresponding refernce for each line.

If we can't parse it as two numbers separated by a period,
just ignore it. There are occasional notes or Coptic page
numbers from the manuscript as well.