Move line reference parse/display into a class.
Encapsulate Drescher's `page.line` style of text reference
in a class, including a `str` representation for printing
and a `from_str` classmethod to parse them out of existing
str representations.

This makes the main reader easier to follow, as well as
simplifying any further manipulations we might want to add.