Print combining marks over a placeholder.
Unicode recommends printing combining marks over the character
U+25CC DOTTED CIRCLE to make it more clear how they interact
with the adjacent character. When printing the character
table for the text, check if the character is marked with a
combining class, and if so prepend a base character.

This doesn't render well with a lot of fonts, which is
unfortunate, but combining with the previous space character
doesn't always give accurate positioning either. I think
we've at least moved the bug onto the fonts and their
interaction with fallback glyphs here.