Generate a markdown table version of the text.
The idea is to have a version of the text that's easy to read
and render through a static site generator. However, Markdown
doesn't really support line numbering for text, which is pretty
important.

We could generate the line numbers with javascript later, but
for now write out a markdown table with explicit line numbering.