Produce html output from the generated markdown.
Use the `markdown` python library to render the markdown table
to html. I'm not convinced this is worthwhile; our text isn't
a great fit for markdown and it's probably just as easy to
generate the html directly.

Attach a header and footer to improve compliance and apply a
minimal stylesheet to make the output more readable.