Add manuscript page numbers.
Update the preparation script to fill in page numbers from the
original manuscript where we have page break marks for that
edition. These are interpolated from the visible Coptic numbers
in the facsimile under the assumption that they form a normal,
linear sequence.

Also correct spacing after removing a *(sic)* annotation at the
beginning or ending of a line.

The sgml document is updated with the output of the new script.