Script to decode Alberto Elli's transliteration
The Mediterraneo Antico series includes a transcription,
gloss, and translation of Hillaria (in Italian) but it
uses an ascii text encoding with a special font a la MdC
for the Coptic. This script is an attempt to convert it
to normalized unicode Coptic script for easier comparison.

E.g. Cut and paste the text from the PDF into a text file
and then run it through this script.